+++ src/main/java/tr/unvercanunlu/ride_share/dto/response/RideRequestedDto.java
@@ -1,5 +1,6 @@
 package tr.unvercanunlu.ride_share.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import tr.unvercanunlu.ride_share.entity.Location;
@@ -10,13 +11,12 @@ public record RideRequestedDto(
     Location pickup,
     Location dropOff,
     double distance,
-    double fare,
+    BigDecimal fare,
     LocalDateTime requestedAt,
     LocalDateTime requestEndAt,
 
     // estimations
     int estimatedDuration
 ) {
-
 
 }
